third_party/blink/web_tests/external/wpt/css/css-pseudo/file-selector-button-inherit-expected.txt

This is a testharness.js-based test.
FAIL ::file-selector-button should inherit from its originating element assert_equals: Check that ::file-selector-button is supported via color expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
